requirements.txt
# TODO: -use a new conda env and install only the required packages
#       -don't use fastapi[all]
alembic==1.4.3
amqp==5.1.1
anyio==3.6.1
arrow==1.2.2
asgiref==3.5.2
atomicwrites==1.4.0
attrs==21.4.0
Authlib==0.14.3
backcall==0.2.0
bcrypt==3.2.0
billiard==3.6.4.0
binaryornot==0.4.4
certifi==2022.5.18.1
cffi==1.15.0
chardet==3.0.4
charset-normalizer==2.0.12
click
click-didyoumean
click-repl
colorama==0.4.4
cryptography==37.0.2
decorator==5.1.1
distlib==0.3.4
dnspython==2.2.1
email-validator==1.2.1
fastapi==0.65.2
filelock==3.7.1
greenlet==1.1.2
h11==0.9.0
httpcore==0.11.1
httptools==0.4.0
httpx==0.15.5
idna==2.10
iniconfig==1.1.1
ipython==7.31.1
itsdangerous==1.1.0
jedi==0.18.1
Jinja2==2.11.3
jinja2-time==0.2.0
kombu==5.2.4
Mako==1.2.0
MarkupSafe==2.1.1
matplotlib-inline==0.1.3
mypy-extensions==0.4.3
orjson==3.7.2
packaging==21.3
parso==0.8.3
passlib==1.7.2
pathspec==0.9.0
pickleshare==0.7.5
platformdirs==2.5.2
pluggy==0.13.1
prompt-toolkit==3.0.29
psycopg2==2.9.3
py==1.11.0
pycparser==2.21
pydantic==1.9.1
Pygments==2.12.0
PyJWT==1.7.1
pyparsing==3.0.9
pytest==7.1.2
python-dateutil==2.8.2
python-dotenv==0.20.0
python-editor==1.0.4
python-multipart==0.0.5
python-slugify==6.1.2
pytz==2022.1
PyYAML==6.0
redis==3.5.3
requests==2.24.0
rfc3986==1.5.0
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.3.19
SQLAlchemy-Utils==0.36.8
starlette==0.14.2
text-unidecode==1.3
toml==0.10.2
tomli==2.0.1
traitlets==5.2.2.post1
typing_extensions==4.2.0
ujson==5.3.0
urllib3==1.25.11
uvicorn==0.12.1
vine==5.0.0
virtualenv==20.14.1
virtualenv-clone==0.5.7
watchgod==0.8.2
wcwidth==0.2.5
websockets==10.3
wincertstore==0.2
